Project Defines: Fix vtx_idx type

+++ Common/Prey/CryCore/Project/ProjectDefines.h
@@ -3,8 +3,8 @@
 #include <Prey/CryCore/BaseTypes.h>
 
 // Uncomment one of the two following typedefs:
-typedef uint32 vtx_idx;
-//typedef uint16 vtx_idx;
+//typedef uint32 vtx_idx;
+typedef uint16 vtx_idx;
 
 // Don't override malloc/free/etc to point to CryEngine allocator
 #define NO_CRY_MALLOC_OVERRIDE 1
